+++ R/rstudio.R
@@ -11,13 +11,19 @@
 #'   * Adds RStudio files to `.Rbuildignore`, if project is a package
 #'
 #' @export
-use_rstudio <- function() {
+#' @param open If `TRUE`, will open project in new RStudio session.
+use_rstudio <- function(open = TRUE) {
   rproj_file <- path_ext_set(project_name(), "Rproj")
   new <- use_template("template.Rproj", rproj_file)
 
   use_git_ignore(".Rproj.user")
   if (is_package()) {
     use_build_ignore(c(rproj_file, ".Rproj.user"))
+  }
+
+  if (open) {
+    done("Opening new project {value(path_file(path))} in RStudio")
+    rstudioapi::openProject(rproj_path(path), newSession = TRUE)
   }
 
   invisible(new)